move protected labels filter
